# specs/001-i-wanto-to/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: Vehicle Diagnostics API
  description: API for Peugeot 2008 vehicle diagnostic data collection and mobile reporting
  version: 1.0.0
  contact:
    name: Vehicle Diagnostics Team
    email: diagnostics@example.com
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.vehiclediagnostics.com/v1
    description: Production server
paths:
  /vehicles:
    get:
      summary: List registered vehicles
      description: Retrieve a paginated list of registered vehicles
      parameters:
        - name: limit
          in: query
          description: Maximum number of vehicles to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of vehicles to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of vehicles retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  vehicles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Vehicle'
                  total_count:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Register a new vehicle
      description: Register a Peugeot 2008 vehicle for diagnostic monitoring
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleCreate'
      responses:
        '201':
          description: Vehicle registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Vehicle with this VIN already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /vehicles/{vehicleId}:
    get:
      summary: Get vehicle details
      description: Retrieve detailed information about a specific vehicle
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vehicle details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    Vehicle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        vin:
          type: string
          pattern: '^[A-HJ-NPR-Z0-9]{17}$'
          description: Vehicle Identification Number (17 characters)
        model_year:
          type: integer
          minimum: 2008
          maximum: 2030
        engine_type:
          type: string
          example: "1.2L PureTech"
        transmission:
          type: string
          enum: ["manual", "automatic", "cvt"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - vin
        - model_year
        - created_at
        - updated_at
    VehicleCreate:
      type: object
      properties:
        vin:
          type: string
          pattern: '^[A-HJ-NPR-Z0-9]{17}$'
        model_year:
          type: integer
          minimum: 2008
          maximum: 2030
        engine_type:
          type: string
        transmission:
          type: string
          enum: ["manual", "automatic", "cvt"]
      required:
        - vin
        - model_year
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
      required:
        - error
        - message
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - ApiKeyAuth: []